+++ packages/studio-base/src/panels/Plot/builders/BlockTopicCursor.ts
