+++ packages/studio-base/src/panels/Plot/csv.ts
@@ -6,7 +6,7 @@ import { Immutable } from "@foxglove/studio";
 import { downloadFiles } from "@foxglove/studio-base/util/download";
 import { formatTimeRaw } from "@foxglove/studio-base/util/time";
 
-import { CsvDataset } from "./IDatasetsBuilder";
+import { CsvDataset } from "./builders/IDatasetsBuilder";
 import { PlotXAxisVal } from "./internalTypes";
 
 function getCSVRow(label: string, data: CsvDataset["data"][0]) {
