+++ packages/studio-base/src/panels/Plot/openSiblingPlotPanel.ts
@@ -4,8 +4,9 @@
 
 import * as _ from "lodash-es";
 
-import type { PlotConfig } from "@foxglove/studio-base/panels/Plot/types";
 import type { OpenSiblingPanel, PanelConfig } from "@foxglove/studio-base/types/panels";
+
+import type { PlotConfig } from "./types";
 
 export function openSiblingPlotPanel(openSiblingPanel: OpenSiblingPanel, topicName: string): void {
   openSiblingPanel({
